+++ CHANGELOG.md
@@ -11,6 +11,12 @@
 
 -->
 
+## 0.10.0 (2023/07/30)
+
+- Change api: T.cond(predicate) -> T.where(condition, parser)
+- Improve types of T.language() API.
+- Improve package build.
+
 ## 0.9.0 (2023/07/21)
 
 - Change: params of T.Parser constructor